// clients/sdk/src/generated/dac-manager/types/taskStatus.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from 'gill';

export enum TaskStatus {
  Ready,
  Pending,
  Processing,
}

export type TaskStatusArgs = TaskStatus;

export function getTaskStatusEncoder(): FixedSizeEncoder<TaskStatusArgs> {
  return getEnumEncoder(TaskStatus);
}

export function getTaskStatusDecoder(): FixedSizeDecoder<TaskStatus> {
  return getEnumDecoder(TaskStatus);
}

export function getTaskStatusCodec(): FixedSizeCodec<
  TaskStatusArgs,
  TaskStatus
> {
  return combineCodec(getTaskStatusEncoder(), getTaskStatusDecoder());
}