// clients/sdk/src/generated/dac-manager/types/computeNodeStatus.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from 'gill';

export enum ComputeNodeStatus {
  Pending,
  Approved,
  Rejected,
  Disabled,
}

export type ComputeNodeStatusArgs = ComputeNodeStatus;

export function getComputeNodeStatusEncoder(): FixedSizeEncoder<ComputeNodeStatusArgs> {
  return getEnumEncoder(ComputeNodeStatus);
}

export function getComputeNodeStatusDecoder(): FixedSizeDecoder<ComputeNodeStatus> {
  return getEnumDecoder(ComputeNodeStatus);
}

export function getComputeNodeStatusCodec(): FixedSizeCodec<
  ComputeNodeStatusArgs,
  ComputeNodeStatus
> {
  return combineCodec(
    getComputeNodeStatusEncoder(),
    getComputeNodeStatusDecoder()
  );
}